one more type fix

+++ limit_order_book/library.py
@@ -27,11 +27,11 @@ def load_library():
     # a type for the unique order ID
     UID = ctypes.c_uint64
     # a type for the quantity of an order
-    Quantity = ctypes.c_uint32
+    Quantity = ctypes.c_int64
     # a type for the price of an order
     Price = ctypes.c_uint64
     # a type for the volume measures in the book
-    Volume = ctypes.c_uint32
+    Volume = ctypes.c_int64
     # a type for counting orders (accumulator)
     Count = ctypes.c_uint32
     # a type for counting orders (accumulator)
